// ui/src/gen/QualityConfiguration.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { QualityStageConfiguration } from "./QualityStageConfiguration";

export type QualityConfiguration = { version: string, project_root: string, stages: Array<QualityStageConfiguration>, };